Initial i3 configuration
This commit configures i3 and the following associated programs:

* picom
* rofi
* i3status

Firefox is configured to be always opened on workspace 10.

The zsh configuration is updated to allow Alacritty and other terminal
emulators to set the window title dynamically. This should not only work
with i3 but also all other window managers / desktop environments.

Additionally this commit contains an initial polybar configuration. I
did not get far with it and currently like i3status better.

The new i3 configuration is not yet integrated into the automated
installation scripts.